backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
numpy==1.26.2
pydantic==2.5.0

# ML Dependencies
ultralytics==8.0.220
torch>=2.0.0
torchvision>=0.15.0
opencv-python==4.8.1.78
scikit-learn==1.3.2
scipy==1.11.4
pandas==2.1.3
matplotlib==3.8.2
seaborn==0.13.0
tqdm==4.66.1
PyYAML==6.0.1
onnx==1.15.0
onnxruntime==1.16.3